backend/requirements.txt
django
gunicorn
pytest
pytest-sugar
pytest-django
django-extensions
django-crispy-forms
crispy-bootstrap5
requests
ruff
psycopg2-binary
bs4
python-dotenv
lxml
# sentence-transformers
